TorrentParser: move StFile struct to separate file and optimize a bit
